auth_service/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
pymongo==4.6.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0

